fix: compile warnings and use generated cmake file (#1)

+++ src/include/nimble-client/game_step_response.h
@@ -8,6 +8,8 @@
 struct NimbleClient;
 struct FldInStream;
 
-int nimbleClientOnGameStepResponse(struct NimbleClient* self, struct FldInStream* inStream);
+#include <stddef.h>
+
+ssize_t nimbleClientOnGameStepResponse(struct NimbleClient* self, struct FldInStream* inStream);
 
 #endif
